requirements.txt
alembic==1.13.1
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
Mako==1.3.2
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==1.26.4
packaging==23.2
passlib==1.7.4
psycopg2==2.9.9
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
scikit-learn==1.2.2
scipy==1.12.0
SQLAlchemy==2.0.26
threadpoolctl==3.2.0
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
